Start generalization to multiple rings
